Snap range end to end of file
We perform validation on range requests to ensure they are entirely
contained in the file, and return HTTP error 416 if they are not. We
return the total file size in responses, so HTTP clients should be able
to ensure their ranges are valid. Unfortunately, some clients request
fixed size ranges without checking.

Handle this by snapping an out-of-bounds range end to the last byte of
the file.